+++ example/android/app/src/main/java/asia/ivity/mediainfoexample/EmbedderV1Activity.java
@@ -2,13 +2,14 @@ package asia.ivity.mediainfoexample;
 
 import android.os.Bundle;
 
+import asia.ivity.mediainfo.MediaInfoPlugin;
 import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
 
-public class MainActivity extends FlutterActivity {
+public class EmbedderV1Activity extends FlutterActivity {
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
+
+    MediaInfoPlugin.registerWith(registrarFor("asia.ivity.mediainfo.MediaInfoPlugin"));
   }
 }